requirements.txt
moviepy==1.0.3
pillow<10
requests
azure-cognitiveservices-speech
python-dotenv
openai
elevenlabs
